Update traing curve and video for merlin_dmlab_collect_good_objects
